cloud_functions/score/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
google-cloud-language
transformers
torch==1.7.1+cpu
torchvision==0.8.2+cpu
torchaudio==0.7.2
bs4